move Builders & Rule into /styles
